Create Day.java
This class creates two arrays of events, one sorted by time and the other sorted by priority, and automatically sorts them into both arrays with the push method. This creates the two views that come with each day of the week- one sorted by the time that it starts, and one sorted by the priority level. There is also a print method that can check and ensure that everything is sorted correctly.