Update ToDo.java
makeTime and calculateEfficiency functions created and begun. There is a makeTime for general cases and for specific time periods. The calculateEfficiency calculates how well inserting a new event at that position would fit into the person’s schedule.

Design Changes: For one, the number and the priority level are directly proportional (the larger the number, the higher the priority), with the lowest priority being zero. Also, the byPriority array list will be sorted with the lowest priority events appearing first.